requirements.txt
flask
gunicorn
azure-functions
azure-ai-textanalytics
python-dotenv
torch
transformers
